Workaround the ActiveRecord / Marshal serialization bug on Ruby 3.2
Fix: https://github.com/mastodon/mastodon/issues/23644

As mentioned on that other thread, the issue arise when a `Status`
is created with SHAPE_TOO_COMPLEX making the instance variable ordering
unpredictable.

But we can sidestep that issue using code inspired from Shopify/paquito.

I added some code in the test environment initializer to cause this
issue to be reproduced and prove this change fixes it.

This code is also easily converted into a monkey patch.